+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.5.0
+- Changed `Pipeline.error` from a property to a method. Call `error()` to re-raise the latest exception, or `error(reraise=False)` to retrieve it without re-raising.
+
 ## 0.4.1
 - Added `Pipeline.__init__()` parameter `run_kwargs` for passing keyword arguments to `run()` when `run_now` is enabled.
 - Added `name` parameter to `Pipeline` for assigning a name to worker threads.